Move two more normalization functions
Move two more normalization functions from the normalization module into
the symbolization logic. Given that they are used only by the latter it
makes much more sense to locate them there.
